Safari-611.1.10/LayoutTests/fast/events/wheel/wheel-event-listeners-on-window-made-passive-expected.txt

Tests that a wheel event handler on the window is passive (cannot prevent scrolling)
PASS windowScrollEventCount > 0 is true
PASS defaultPrevented is false
PASS successfullyParsed is true

TEST COMPLETE
